Add plugin to default config; refactor exports
