+++ src/javascript/CKEditor/Picker/InsertJahiaImage/index.js
@@ -1,0 +1,1 @@
+export {InsertJahiaImage} from './InsertJahiaImage';
