+++ src/javascript/CKEditor/Picker/index.js
@@ -1,0 +1,3 @@
+export * from './InsertJahiaImage';
+export {Picker} from './Picker';
+export {JahiaLinkProvider} from './JahiaLinkProvider';
