--- src/javascript/CKEditor/Picker/plugins/index.js
@@ -1,1 +1,0 @@
-export * from './InsertJahiaImage/InsertJahiaImage';
